+++ packages/expo-router/src/split-view/index.ts
@@ -1,3 +1,4 @@
 export { SplitView, type SplitViewProps } from './split-view';
 export * from './elements';
+export { setSplitViewImplementation, type SplitViewImplementation } from './implementation';
 export type { SplitHostProps } from 'react-native-screens/experimental';